자신의 글만 삭제

+++ backend/src/main/java/com/example/backend/controller/board/BoardController.java
@@ -3,8 +3,9 @@ package com.example.backend.controller.board;
 import com.example.backend.dto.board.Board;
 import com.example.backend.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
-import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
@@ -36,13 +37,21 @@ public class BoardController {
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity<Map<String, Object>> delete(@PathVariable int id) {
-        if (service.remove(id)) {
-            return ResponseEntity.ok().body(Map.of("message", Map.of("type", "success",
-                    "text", STR."\{id}번 게시글이 삭제되었습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> delete(@PathVariable int id,
+                                                      Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            if (service.remove(id)) {
+                return ResponseEntity.ok().body(Map.of("message", Map.of("type", "success",
+                        "text", STR."\{id}번 게시글이 삭제되었습니다.")));
+            } else {
+                return ResponseEntity.internalServerError().body(Map.of("message", Map.of("type", "error",
+                        "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+            }
         } else {
-            return ResponseEntity.internalServerError().body(Map.of("message", Map.of("type", "error",
-                    "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+            return ResponseEntity.status(403)
+                    .body(Map.of("message", Map.of("type", "error",
+                            "text", "삭제 권한이 없습니다.")));
         }
     }
 
@@ -61,6 +70,7 @@ public class BoardController {
     }
 
     @PostMapping("add")
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<Map<String, Object>> add(@RequestBody Board board,
                                                    Authentication authentication) {
